test(cli): split json capture end-to-end suite
Move the no-trace JSON journal contract into a dedicated binary suite so resume coverage stays cohesive and every Rust file remains below the enforced 1,500-line ceiling.
